Multiple Tile Movement started, but broken
I put in a lot of bug-testing and console outputs to find errors.

It works for moving South and West.

The problem with calling the moveAdjacent function repeatedly is that a
lot of the moves are returning false and thus not executing.

The first moves are executing, but the successive moves aren’t,
presumably because some property isn’t being updated when the creature
moves one tile, resulting in failure for the rest of the tile movements
because the information being passed as a parameter about the creature
is out of date.