Attackable spots are only attackable if they are within range
We may want to change this later by making it so you can move to an
attackable tile and then attack it but at the moment this hasn’t been
done, because that would require the game to tell you how it wants you
to move. The game would have to choose a location around the attackable
enemy, but it would have to tell you which side it was going to.